Refactoring local flask to operate now given a json with trial permuations rather than doing randomization itself
