--- frontend/public/demo2.json
@@ -1,21 +1,0 @@
-{
-  "studyName": "asdfasdf",
-  "studyDescription": "",
-  "studyDesignType": "Between",
-  "participantCount": "23",
-  "tasks": [
-    {
-      "taskName": "asdf",
-      "taskDescription": "",
-      "taskDirections": "",
-      "taskDuration": "",
-      "measurementOptions": []
-    }
-  ],
-  "factors": [
-    {
-      "factorName": "asdfasdf",
-      "factorDescription": ""
-    }
-  ]
-}
