--- frontend/public/demo3.json
@@ -1,42 +1,0 @@
-{
-    "factors": [
-        {
-            "factorDescription": "Factor Des",
-            "factorID": 21,
-            "factorName": "USB"
-        },
-        {
-            "factorDescription": "bru",
-            "factorID": 22,
-            "factorName": "Laptop"
-        } 
-    ],
-    "participantCount": "214",
-    "studyDescription": "Study Desc",
-    "studyDesignType": "Between",
-    "studyName": "Study Name",
-    "tasks": [
-        {
-            "measurementOptions": [
-                "Mouse Movement"
-            ],
-            "taskDescription": "Task Des",
-            "taskDirections": "Move your mouse foo",
-            "taskDuration": "5.000",
-            "taskID": 21,
-            "taskName": "Mouse"
-        },
-        {
-            "measurementOptions": [
-                "Keyboard Inputs"
-            ],
-            "taskDescription": "Task Des",
-            "taskDirections": "Start spamming keyboard lil bru",
-            "taskDuration": "5.000",
-            "taskID": 22,
-            "taskName": "keyboard"
-        }
-    ],
-    "participantSessId": 9,
-    "study_id": "11"
-}
