+++ local_backend/utility/file_management.py
@@ -1,0 +1,14 @@
+# Purpose: functionality for creating the zip files for each session
+
+
+import os
+import zipfile
+
+
+def zip_folder(folder_path, zip_name):
+    with zipfile.ZipFile(zip_name, 'w', zipfile.ZIP_DEFLATED) as zipf:
+        for root, dirs, files in os.walk(folder_path):
+            for file in files:
+                file_path = os.path.join(root, file)
+                arcname = os.path.relpath(file_path, folder_path)
+                zipf.write(file_path, arcname)
